storage-types: factor out StringOrSecret into a module
To avoid a circular dependency in a forthcoming commit between AWS
connections and Kafka connections.
