+++ src/storage-types/src/connections.proto
@@ -15,30 +15,25 @@ import "repr/src/global_id.proto";
 import "repr/src/url.proto";
 import "proto/src/tokio_postgres.proto";
 import "storage-types/src/errors.proto";
+import "storage-types/src/connections/string_or_secret.proto";
 
 package mz_storage_types.connections;
 
-message ProtoStringOrSecret {
-    oneof kind {
-        string string = 1;
-        mz_repr.global_id.ProtoGlobalId secret = 2;
-    }
-}
 
 message ProtoTlsIdentity {
-    ProtoStringOrSecret cert = 1;
+    string_or_secret.ProtoStringOrSecret cert = 1;
     mz_repr.global_id.ProtoGlobalId key = 2;
 }
 
 message ProtoKafkaConnectionTlsConfig {
-    ProtoStringOrSecret root_cert = 1;
+    string_or_secret.ProtoStringOrSecret root_cert = 1;
     ProtoTlsIdentity identity = 2;
 }
 
 message ProtoKafkaConnectionSaslConfig {
     reserved 4;
     string mechanism = 1;
-    ProtoStringOrSecret username = 2;
+    string_or_secret.ProtoStringOrSecret username = 2;
     mz_repr.global_id.ProtoGlobalId password = 3;
 }
 
@@ -60,21 +55,21 @@ message ProtoKafkaConnection {
     ProtoTunnel default_tunnel = 7;
     optional string progress_topic = 5;
     ProtoKafkaTopicOptions progress_topic_options = 10;
-    map<string, mz_storage_types.connections.ProtoStringOrSecret> options = 6;
+    map<string, mz_storage_types.connections.string_or_secret.ProtoStringOrSecret> options = 6;
     ProtoKafkaConnectionTlsConfig tls = 8;
     ProtoKafkaConnectionSaslConfig sasl = 9;
 }
 
 message ProtoCsrConnection {
     mz_repr.url.ProtoUrl url = 1;
-    ProtoStringOrSecret tls_root_cert = 2;
+    string_or_secret.ProtoStringOrSecret tls_root_cert = 2;
     ProtoTlsIdentity tls_identity = 3;
     ProtoCsrConnectionHttpAuth http_auth = 4;
     ProtoTunnel tunnel = 6;
 }
 
 message ProtoCsrConnectionHttpAuth {
-    ProtoStringOrSecret username = 1;
+    string_or_secret.ProtoStringOrSecret username = 1;
     mz_repr.global_id.ProtoGlobalId password = 2;
 }
 
@@ -82,10 +77,10 @@ message ProtoPostgresConnection {
     string host = 1;
     uint32 port = 2;
     string database = 3;
-    ProtoStringOrSecret user = 4;
+    string_or_secret.ProtoStringOrSecret user = 4;
     mz_repr.global_id.ProtoGlobalId password = 5;
     mz_proto.tokio_postgres.ProtoSslMode tls_mode = 6;
-    ProtoStringOrSecret tls_root_cert = 7;
+    string_or_secret.ProtoStringOrSecret tls_root_cert = 7;
     ProtoTlsIdentity tls_identity = 8;
     ProtoTunnel tunnel = 12;
 }
@@ -128,10 +123,10 @@ enum ProtoMySqlSslMode {
 message ProtoMySqlConnection {
     string host = 1;
     uint32 port = 2;
-    ProtoStringOrSecret user = 3;
+    string_or_secret.ProtoStringOrSecret user = 3;
     mz_repr.global_id.ProtoGlobalId password = 4;
     ProtoMySqlSslMode tls_mode = 5;
-    ProtoStringOrSecret tls_root_cert = 6;
+    string_or_secret.ProtoStringOrSecret tls_root_cert = 6;
     ProtoTlsIdentity tls_identity = 7;
     ProtoTunnel tunnel = 8;
 }
