+++ src/storage-types/src/connections.rs
@@ -52,6 +52,7 @@ use url::Url;
 
 use crate::configuration::StorageConfiguration;
 use crate::connections::aws::{AwsConnection, AwsConnectionValidationError};
+use crate::connections::string_or_secret::StringOrSecret;
 use crate::controller::AlterError;
 use crate::dyncfgs::{
     ENFORCE_EXTERNAL_ADDRESSES, KAFKA_CLIENT_ID_ENRICHMENT_RULES,
@@ -62,6 +63,7 @@ use crate::AlterCompatible;
 
 pub mod aws;
 pub mod inline;
+pub mod string_or_secret;
 
 include!(concat!(env!("OUT_DIR"), "/mz_storage_types.connections.rs"));
 
@@ -104,72 +106,6 @@ impl SecretsReaderExt for Arc<dyn SecretsReader> {
         async move { sr.read_string(id).await }
             .run_in_task_if(in_task, || "secrets_reader_read".to_string())
             .await
-    }
-}
-
-#[derive(Arbitrary, Clone, Debug, Eq, PartialEq, Hash, Serialize, Deserialize)]
-pub enum StringOrSecret {
-    String(String),
-    Secret(GlobalId),
-}
-
-impl StringOrSecret {
-    /// Gets the value as a string, reading the secret if necessary.
-    pub async fn get_string(
-        &self,
-        in_task: InTask,
-        secrets_reader: &Arc<dyn SecretsReader>,
-    ) -> anyhow::Result<String> {
-        match self {
-            StringOrSecret::String(s) => Ok(s.clone()),
-            StringOrSecret::Secret(id) => secrets_reader.read_string_in_task_if(in_task, *id).await,
-        }
-    }
-
-    /// Asserts that this string or secret is a string and returns its contents.
-    pub fn unwrap_string(&self) -> &str {
-        match self {
-            StringOrSecret::String(s) => s,
-            StringOrSecret::Secret(_) => panic!("StringOrSecret::unwrap_string called on a secret"),
-        }
-    }
-
-    /// Asserts that this string or secret is a secret and returns its global
-    /// ID.
-    pub fn unwrap_secret(&self) -> GlobalId {
-        match self {
-            StringOrSecret::String(_) => panic!("StringOrSecret::unwrap_secret called on a string"),
-            StringOrSecret::Secret(id) => *id,
-        }
-    }
-}
-
-impl RustType<ProtoStringOrSecret> for StringOrSecret {
-    fn into_proto(&self) -> ProtoStringOrSecret {
-        use proto_string_or_secret::Kind;
-        ProtoStringOrSecret {
-            kind: Some(match self {
-                StringOrSecret::String(s) => Kind::String(s.clone()),
-                StringOrSecret::Secret(id) => Kind::Secret(id.into_proto()),
-            }),
-        }
-    }
-
-    fn from_proto(proto: ProtoStringOrSecret) -> Result<Self, TryFromProtoError> {
-        use proto_string_or_secret::Kind;
-        let kind = proto
-            .kind
-            .ok_or_else(|| TryFromProtoError::missing_field("ProtoStringOrSecret::kind"))?;
-        Ok(match kind {
-            Kind::String(s) => StringOrSecret::String(s),
-            Kind::Secret(id) => StringOrSecret::Secret(GlobalId::from_proto(id)?),
-        })
-    }
-}
-
-impl<V: std::fmt::Display> From<V> for StringOrSecret {
-    fn from(v: V) -> StringOrSecret {
-        StringOrSecret::String(format!("{}", v))
     }
 }
