+++ src/storage-types/src/connections/string_or_secret.proto
@@ -1,0 +1,23 @@
+// Copyright Materialize, Inc. and contributors. All rights reserved.
+//
+// Use of this software is governed by the Business Source License
+// included in the LICENSE file.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0.
+
+// buf breaking: ignore (currently does not require backward-compatibility)
+
+syntax = "proto3";
+
+import "repr/src/global_id.proto";
+
+package mz_storage_types.connections.string_or_secret;
+
+message ProtoStringOrSecret {
+    oneof kind {
+        string string = 1;
+        mz_repr.global_id.ProtoGlobalId secret = 2;
+    }
+}
